third_party/WebKit/LayoutTests/fast/events/hit-test-cache-iframes-expected.txt

Ensure hit test cache is cleared across iframes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.hitTestCacheHits(document) is 0
PASS window.internals.hitTestCacheHits(document) is 0
PASS window.internals.hitTestCacheHits(document) is 0
PASS successfullyParsed is true

TEST COMPLETE
